+++ update-modules.sh
@@ -61,7 +61,7 @@ IFS=';'
 
 cat $mainDir/$modulesfilename | while read -A repo; do
 	echo "found github repos to clone $repo[1] $repo[2] $repo[3]"	
-	cd $envDir
+	cd $envDiro
 	if [ ! -d $envDir/$repo[2] ]; then
 		git clone $repo[1] $repo[2];
 		cd $envDir/$repo[2]; 
@@ -73,8 +73,12 @@ cat $mainDir/$modulesfilename | while read -A repo; do
 	cd $envDir/$repo[2]; 
 
 	#update repo
-	git fetch;git pull;
-	sed -i '' 's/7\.0\.0\.0/7\.3\.0\.0/g' pom.xml
-	#deploy repo to profile
-	mvn -P $profile -D skipTests clean install jahia:deploy;
+	updated=$(git fetch;git pull)
+	if [[ $updated != "Already up to date." ]]; then
+		sed -i '' 's/7\.0\.0\.0/7\.3\.0\.0/g' pom.xml
+		#deploy repo to profile
+		mvn -P $profile -D skipTests clean install jahia:deploy;
+	else 
+		echo $updated	
+	fi
 done
